move the visibility header
